Wrap all code into functions
Wrap all code in functions.  Use the name
as a trigger for actual execution.

For slightly improved speed, we actually
define `shpec` as a function.

This allows us to run shpec as a script or
as a function.  Naturally, the script just
calls the function.

To protect the environment all processing
occurs inside a subshell.