Remove unneeded assertion
We cannot stub `exit` in every POSIX shell.
In ksh at least `exit` is a special built-in and can not
be redefined as a function.

We use `false` instead.

This delta removes a pre-stub test verifying the normal behaviour
of `false`. What matters is that the stub successfully replaces
the original, and the unstub restores the default.