Add skip support
Adds function `skip`.  If a test invokes `skip`
prior to an assertion, the assertion will be
resolved, but failing match will not result in a failure,
and instead count as a *skipped* test.

This update adds counters for passed and skipped tests.
This also changes the summary line, and color codes.

A third color is introduced for skipped tests: yellow.

The summary line will report numbers in black, with
the words 'failures', 'passed', and 'skipped' in
red, green, and yellow, respectively.

Finally, tests are included for the presence of
a terminal.  If no terminal is present, color codes
are replaced as follows:

green   OK      passed
red     NO      failed
yellow  ??      skipped