Fix assertion-grouping tests for BSD userland
Use our 'match' matcher to simplify our assertions. It forces us to be
quite liberal with wildcards and escaping special characters (and, most
frustratingly, whitespace), but it works on both BSD and GNU.

The "run in a subshell and parse out a bunch of special characters"
method was not portable, and I did not feel it was necessarily for the
tests.