WIP: Remove default wildcards from glob matcher
Using the multi-character wildcard to surround the matcher argument
seems likely to lead to unexpected behavior, mainly in the form of false
positives due to the leniency/greediness of said multi-character
wildcards.

This removes that behavior, creates a dedicated test section for the
glob matcher, and adds a few more tests documenting the new behavior.

NB that a lot of shpec's tests were relying on this greedy, lenient
behavior; this commit is a WIP because all of those tests are now
broken.