Properly handle arguments with whitespace
We have two distinct situations here:

1. We were given no arguments
2. We were given arguments

The first case has always worked; we spawn a subshell to find our test
files, assign it to a variable, and then iterate over them to source/run
them.

The second case is a bit more complicated: to get our arguments as they
were originally quoted, we need to use `"$@"`. However, we cannot assign
this to a variable, lest it become a single string again.

All that to say we can no longer conditionally assign and then loop over
that variable; we now need to have to separate loops to handle these two
distinct situations.

But it works!

Closes #112.